+++ src/util/is-browser.js
@@ -1,0 +1,3 @@
+module.exports = function isBrowser () {
+  return (typeof window !== 'undefined')
+}
