+++ src/util/is-node.js
@@ -1,0 +1,6 @@
+module.exports = function isNode () {
+  return (
+    typeof module !== 'undefined' &&
+    typeof module.exports !== 'undefined'
+  )
+}
