+++ src/util/to-array.js
@@ -1,0 +1,3 @@
+module.exports = function toArray (arg) {
+  return Array.prototype.slice.call(arg, 0)
+}
