+++ test/karma.conf.js
@@ -23,7 +23,7 @@ var conf = {
   // start these browsers
   // available browser launchers: https://npmjs.org/browse/keyword/karma-
   // launcher
-  browsers: ['Chrome', 'Firefox'],
+  browsers: ['Chrome'],
 
   singleRun: false
 }
